According to the CreateDBInstance documentation, the parameter PerformanceInsightsKMSKeyId can be either a KMS key ARN, a KMS alias ARN or simply a KMS key ID.
In most cases, I would expect people to rely on either the KMS alias ARN or the key ARN, but because of the way data.aws_kms_key works, some users might end up with a raw key ID (not the ARN).

In this case the cluster instance creation will fail as the validation ensures that the performance_insights_kms_key_id is an ARN.

As a side note, the terraform documentation also mentions that we expect "The ARN for the KMS key to encrypt Performance Insights data.", it is probably worth specifying that an ARN to an alias is perfectly fine as well.

Fixing https://github.com/terraform-providers/terraform-provider-aws/issues/3014