Store all config-specified DB params in state.
Prior to this commit, only database parameters that AWS recognised as
being user-altered were stored in state. This would be fine, except if
users specified the same value for a parameter as AWS' default for that
value, they would get a perpetual diff, because AWS would report it as a
non-user-changed parameter. Meaning it wouldn't be set in state, no
matter how many times the user applied.

This fixes the problem by retrieving all parameters, regardless of how
AWS reports them being set. We then select out from those parameters
only the ones that report that they're user-modified _or_ those that are
found in a config, if a config is available.

This commit also adds a test which exhibits this behaviour, which failed
before this fix was applied, and passes now that it has been applied.