r/lb_target_group: Allow for stickiness config on TCP groups (continued)
This is a continuation of 08d7b8c, which modifies the diff validation
check to allow for stickiness to be defined on TCP target groups, but
requires that this setting be false. While this allows the diff to be
applied, this still fails on the AWS end because it still sets the
stickiness attributes on target group types that don't support it.

This update skips sending the stickiness attribute data to AWS when
using a TCP load balancer, in addition to skipping reading it back in
when the real-world protocol is TCP. This ensures both that the apply
will succeed in AWS, and that spurious diffs aren't caused by ensuring
that the state coming from config is not overwritten.

Also, tests have been added for both ensuring that enabling stickiness
is still blocked, and ensuring that a disabled stickiness block
succeeds with an empty post-apply plan.

Finally, documentation has been added denoting that you can still
declare a stickiness block when TCP is used, but it must be disabled.