Update docs to reflect API behavior
The documentation stated that `delete_on_termination` field for each `network_interface` could be set to `true` or `false`. However the behavior of the API does not allow this, specifically the API will respond with an error if `delete_on_termination` is set to true with a `network_interface_id` being specified:

Currently it is documented that `network_interface_id` is a required field, which implies that there is no way to successfully apply with `delete_on_termination` set to true. Therefore the line of documentation for `delete_on_termination` is not valid.