Adding processing_configuration to Elasticsearch section
This is a partial fix for the issue listed below. The problem currently is
that you can add an extended_s3_configuration to the
aws_kinesis_firehose_delivery_stream and set the destionation to 'extended_s3'
if you want data transformation to be enabled for S3, but you can't do the
same if your destination is 'elastisearch' - this commit fixes that.

Support Kinesis Firehose Data Transformation #513
https://github.com/terraform-providers/terraform-provider-aws/issues/513