Improve PR #1587 and with that fix issue #2069
The logic in PR #1587 has some flaws with regards to how the security group API works in relation with TF.

The problem will not show itself when you manage everything properly from TF, but when someone makes a manual change directly in AWS you could be hit by the problem described in issue #2069.

This PR aims to fix that problem. As for the code itself, it does look like it’s not really DRY, but if you look closely the code is slightly different every time. So making an abstraction seems questionable in this case…