Update cognito docs for verification by link
If you use `{####}` in `email_message_by_link` you're returned an ambiguous message from the API: "the message is empty". This is because the placeholder needs to be, specifically, `{##VERIFY EMAIL##}`. I'm not clear on why the regexp in the validation* funcs do not catch this, and currently do not have the time to dig into that. However I figured I'd submit a doc improvement to save other folks time on this one.