aws: add pagination support to the findRecord func for route53
The ListResourceRecordSets API call may return ALIAS records that point
to the record specified by the "name" parameter. These records may
appear first because the results are lexicographically ordered. When
there are more than 100 of these records they can cause the target
record to be part of the truncated set.

Check for a truncated indicator in the ListResourceRecordSetsOutput. If
the response is truncated, repeat the API call with an updated starting
name, type, and identifier. If the response is not truncated, return an
r53NoRecordsFound.