Add contraints link for aws_db_instance attribute
The attribute db_subnet_group_name for resource aws_db_instance should not be
specified if read replicas are in the same region as the master. This change adds
a clarification in the docs.

Sample error from Terraform v0.11.5, provider.aws v1.11.0:
Error creating DB Instance: DBSubnetGroupNotAllowedFault: DbSubnetGroupName
should not be specified for read replicas that are created in the same region as the master

Moreover, the attribute has a number of constraints and this change adds a link to the AWS documentation.