Drop not-working cloudtrail example
Logging to s3 is the only option for the cloudtrail resource. So the current basic example doesn't work, and incorrectly implies that it's possible to log to someplace other than s3. This patch drops the basic example and relabels the s3 example as basic, since that's the simplest working configuration.