Improve handling of security group updates
Currently, when any single rule is added or removed to/from a security group,
Terraform will revoke and re-authorize all rules in the given ingress/egress
block. This leads to a couple of undesirable effects:

1. Rules that didn't change are always removed and re-added back on apply,
   causing temporary packet loss, and potentially - outage.
2. Due to EC2 API specifics, if any single rule in the given ingress/egress
   block cannot be authorized because of an error (e.g. wrong netmask,
   non-existent security group, etc), no rules are added at all after revoke,
   and the security group may be left blank causing permanent packet loss,
   outage, and manual intervention is then required to recover.

This patch improves the way security groups are updated so that only rules that
actually changed, are authorized/revoked, which avoids unnecessary revocation
of entire ingress/egress blocks and reduces the possibility of an outage.