Document that the protocols available to a listener are determined by the load balancer type
The docs don't make it particularly clear right now that 'TCP' protocol listeners are only available on Network Load Balancers and vice versa.

This isn't something we can catch at plan or even apply time so have to just rely on the AWS API telling the user that they can't use the wrong protocol for the wrong load balancer type.

Closes #4659.