Link to Aurora docs to explain the difference between Aurora MySQL 1.x and 2.x
The differences between Aurora MySQL 1.x and Aurora MySQL 2.x aren't the clearest so it helps to link to the docs in places where people are making this decision.

For reference Aurora MySQL 1.x is based on the MySQL 5.6.x engine (and is wire compatible with it) with some specific improvements.
Aurora MySQL 2.x is currently based on the MySQL 5.7.x engine (and is wire compatible with it) with some specific improvements that are not all the same as the improvements made in Aurora MySQL 1.x.

It also doesn't help that AWS is continuing to develop functionality for Aurora MySQL 1.x separately to Aurora MySQL 2.x so these are currently divering slightly.