Fix the access policy in the ES domain docs
If a resource isn't specified AWS automatically adds it to point to the ES domain it has been created for.
This causes a perpetual diff where Terraform keeps attempting to remove the Resource key and AWS keeps adding it back.

Also adds a note drawing attention to the fact that the values for the advanced options _must_  be strings or this will also cause a perpetual diff.