Raise an error when failing to retrieve RDS instance tags
Just logging the error means that if the user is unable to call the ListTagsForResource API endpoint then they will see a diff showing that they need to add the tags each time.
Instead it's better to actually error instead of hiding this from the user (unless they happen to check the logs).

Closes https://github.com/terraform-providers/terraform-provider-aws/issues/345