Force new route when changing the route table
Before this change, changing the route table of a route showed an in place modification for Terraform to perform.
AWS actually creates a new route in the new route table with the ReplaceRoute API call but this left the old route lying around and no longer managed.

The acceptance test provided in this commit was failing before because Terraform was attempting to update the new route's 10.3.0.0/16 route to point to the new target but the 10.3.0.0/16 route didn't exist in that route table and instead wanted Terraform to call the CreateRoute API instead:

> There is no route defined for '10.3.0.0/16' in the route table. Use CreateRoute instead.

Instead we should just use ForceNew to make Terraform create a new route in the new route table and then delete the old one.