Ignore CPU credit specification when not using a T2 type
If the 'CreditSpecificationRequest' struct is provided at all then the AWS API will throw an error saying that the instance type does not support T2 unlimited even if it is set to 'standard'.

Instead we check if the instance type is 't2' before adding it to the instance opts and if not just warn that it was specified but then ignored.

Fixes #5067.