Fix read for non-pending aws_acm_certificate
If a DomainValidation indicates a validation status of success or failed
there is no reason for it to be re-queried, even if no ResourceRecord or
ValidationEmails are present. Re-querying in either of these cases will
cause the import of the resource to fail and prevent the certificate
from being imported.

Now the certificate is only re-queried if the DomainValidation is in a
pending state.