tests: Implement testAccOrganizationsAccountPreCheck function and add to existing AWS Organizations based acceptance tests
When testing resources that involve AWS Organizations interaction, usually the acceptance testing needs the ability to create and delete an AWS Organization. When an account is already an existing member of an AWS Organization (e.g. for consolidated billing), it will expectedly throw errors such as these:

--- FAIL: TestAccAWSConfigConfigurationAggregator_organization (7.08s)
	testing.go:518: Step 0 error: Error applying: 2 errors occurred:
			* aws_organizations_organization.test: 1 error occurred:
			* aws_organizations_organization.test: Error creating organization: AlreadyInOrganizationException: The AWS account is already a member of an organization.

Here we introduce a PreCheck function that ensures the AWS account running the acceptance testing is explicitly not a member of an organization already.

Verified using an account that is a member of an organization:

=== RUN   TestAccAWSConfigConfigurationAggregator_organization
--- SKIP: TestAccAWSConfigConfigurationAggregator_organization (1.37s)
	provider_test.go:132: skipping tests; this AWS account must not be an existing member of an AWS Organization

As well as one not a member of an organization:

=== RUN   TestAccAWSConfigConfigurationAggregator_organization
--- PASS: TestAccAWSConfigConfigurationAggregator_organization (15.72s)