Minor bug fixes and adds acceptance test step to ensure updating domain of db instance works as expected.
This commit integrates suggesting made by @bflad in review.  It also merges any updates to domain or domain_iam_role_name into a single update request, as both arguments must be present in order to update either property of a DB instance.  Finally, it adds an accpetance test step to ensure that after a db instance is joined to a domain, it can be succesfully moved to another by updating the domain argument.X