Data source aws_db_snapshot: Shared and public snapshot fix
A combination of `snapshot_type = "shared"` and `db_instance_identifier`
would always return an empty result set due to the server side filtering
against the DB Instance Identifier only considering instance identifiers
from the local account, as opposed to from the source account.

It is still possible to filter the results client-side, so do that
instead.