Adding a note to help people avoid a common mistake
I'm sure there have been plenty of people who struggled with this, I wasted a few hours on it myself. Might be helpful to put a note in to help save people the suffering. The problem is described here: https://github.com/hashicorp/terraform/issues/16285, but it would be easier to not have to find that in the first place.