resource/aws_redshift_cluster: Properly disable logging when using logging nested argument
This is a quick fix until the deprecated `enabled_logging`, etc. arguments are removed in the next major version. Previously, when switching from logging being enabled to disabled, the update function would errantly reference both the deprecated `enable_logging` value from state (which would be true) and the new value `logging.0.enable` from configuration (which would be false) with a logical OR to determine whether logging should be enabled or disabled. We now treat the handling between the logging and deprecated arguments separately.

The reason this was not previously handled was that the existing acceptance testing had a race condition that meant the EnableLogging call would not error if DeleteBucket had not been called yet, but otherwise could fail with:

```
--- FAIL: TestAccAWSRedshiftCluster_loggingEnabled (1157.85s)
    testing.go:527: Step 1 error: Error applying: 1 error occurred:
        	* aws_redshift_cluster.default: 1 error occurred:
        	* aws_redshift_cluster.default: BucketNotFoundFault: Could not find bucket with name tf-redshift-logging-5762787809933873609
```

Now the testing passes and verified that `DisableLogging` is correctly called in both tests:

```
--- PASS: TestAccAWSRedshiftCluster_loggingEnabled (1377.47s)
--- PASS: TestAccAWSRedshiftCluster_loggingEnabledDeprecated (1587.61s)
```