elasticsearch_domain: Add MaxItems=1 annotations
In the `aws_elasticsearch_domain` resource, the three blocks for
`cluster_config`, `ebs_options` and `snapshot_options` are described as
lists with nested schema, however only one element is permitted. This
was previously enforced by custom code in the Create and Update
functions.

This commit makes use of the `MaxItems: 1` facility of `helper/schema`
in order to remove custom code and produce better error messages in the
face of non-compliance with the reuqirements of the schema.