s3_bucket: Add MaxItems=1 annotations
In the `aws_s3_bucket` resource, the `website` block for is described as
a list with nested schema, however only one element is permitted. This
was previously enforced by code in the `resourceAwsS3BucketWebsiteUpdate`
function.

This commit makes use of the `MaxItems: 1` facility of `helper/schema`
in order to remove custom code and produce better error messages in the
face of non-compliance with the reuqirements of the schema.