Fix bad advice on using ".name" property of Role while ignoring path
- When creating an IAM Role with a non-default path (anything other than
"/"), it is not sufficient to use only the "name" property of the Role
when referring to it from the Glue Crawler resource (and probably
others as well). Therefore, in the examples I have replaced usage of
"name" with usage of "arn", and I updated the documentation on the
"role" argument of the aws_glue_crawler resource to clarify the meaning
of "The IAM role" which was not well-defined.
- Relevant Terraform version: This update can be deployed to the site
immediately, it is not specific to an upcoming version of Terraform.