Allow keeping the old task definition revisions when updating
Task definition revisions are immutable so Terraform is unable to just update this resource and instead needs to delete the old revision and create a new one.
This isn't helpful if you still want access to the old task definition such as if you are using another system to manage rollbacks etc as you are unable to roll a service back to an inactive task definition.

Closes https://github.com/terraform-providers/terraform-provider-aws/issues/258