resource/aws_flow_log: Automatically trim :* suffix from log_destination argument
`aws_cloudwatch_log_group` `arn` attribute references will automatically contain `:*` due to the CloudWatch Logs API implementation. While we may in the future trim that attribute in the `aws_cloudwatch_log_group` resource in a major release, this provides a quick band aid for issues when using it with this resource.

Changes:
* resource/aws_flow_log: Automatically trim `:*` suffix from `log_destination` argument
* tests/resource/aws_flow_log: Switch acceptance testing to `resource.ParallelTest()`

```
--- PASS: TestAccAWSFlowLog_LogDestinationType_CloudWatchLogs (25.45s)
--- PASS: TestAccAWSFlowLog_LogDestinationType_S3 (28.20s)
--- PASS: TestAccAWSFlowLog_SubnetID (30.30s)
--- PASS: TestAccAWSFlowLog_VPCID (41.48s)
```