Closes #6399: Adds an aws_iam_assume_role_policy resource
This adds a resource that can be used to update the assume role
policy for an IAM role created outside of terraform, without the
need to import that role into terraform state manually.

This is useful in cases where a team using terraform is allowed
to manage permissions for IAM roles, but does not have access to
create or delete IAM roles.