tests/resource/aws_ami_copy: Simplify basic test and add description test
The usage of the `aws_ami_from_instance` resource was flakey at best and highlighted an issue that needs to be addressed in a separate bug fix. Daily acceptance testing is reporting this currently with `TestAccAWSAMICopy_basic`: Frequent test status changes: 9 changes out of 66 invocations

With the following error:

```
--- FAIL: TestAccAWSAMICopy_basic (58.68s)
    testing.go:538: Step 0 error: Error applying: 1 error occurred:
          * aws_ami_from_instance.test: 1 error occurred:
          * aws_ami_from_instance.test: Error waiting for AMI (ami-07772720b881c21d9) to be ready: unexpected state 'failed', wanted target 'available'. last error: %!s(<nil>)
```

As an added bonus, this reduces the runtime of `TestAccAWSAMICopy_basic` by a minute or two.

Changes:
* tests/resource/aws_ami_copy: Switch `TestAccAWSAMICopy_basic` from `aws_ami_from_instance` resource usage to directly using `aws_ebs_volume`, `aws_ebs_snapshot`, and `aws_ami` resources
* tests/resource/aws_ami_copy: Add `TestAccAWSAMICopy_Description` acceptance test for verifying `description` attribute on creation and updates

Output from acceptance testing:

```
--- PASS: TestAccAWSAMICopy_EnaSupport (368.45s)
--- PASS: TestAccAWSAMICopy_basic (394.48s)
--- PASS: TestAccAWSAMICopy_Description (405.82s)
```