Make tags truly optional on the transfer sftp server resource
The aws transfer sftp server resource was introduced in
https://github.com/terraform-providers/terraform-provider-aws/pull/6639
and it's wonderful.

In using it, i noticed that the documentation says that the tags are
optional, but if they're not provided then the AWS API responds with an
error. I investigated and the tags are truly optional on AWS's side, but
the code (in its current form) sets the `Tags` attribute to `[]` if the
resource is not defined with tags. This results in an invalid API
request to AWS. Here's an example of the error output:

```
Error: Error applying plan:

1 error(s) occurred:

* aws_transfer_server.sftp_server: 1 error(s) occurred:

* aws_transfer_server.sftp_server: Error creating Transfer Server: InvalidParameter: 1 validation error(s) found.
- minimum field size of 1, CreateServerInput.Tags.
```

This changeset only sets the tags when they're not empty. Voila!