service/vpc: Fix dangling resources in DHCP Options Set data source test
TestAccDataSourceAwsVpcDhcpOptions_Filter was previously ending with a
configuration containing errors. In the 0.11-sourced SDK this is for some
reason silently ignored, with the test leaving behind a dangling
DHCP options set in the target account. In the 0.12-sourced SDK this
became an error as expected, because running "terraform destroy" with an
erroneous config is normally blocked, and so this test would've begun
failing after SDK upgrade.

By adding an additional step with an empty config here, we behave as if
the user responded to the error in the previous step by deleting all of
the resource blocks from the configuration and trying "terraform apply"
again. That final step then succeeds, in the process destroying the
resources created in previous steps, and allowing the test to conclude
with a valid configuration and empty state.