Fix GH-7397: Ensure subnet_mapping is present if configured
Issue: GH-7397 (See: https://github.com/terraform-providers/terraform-provider-aws/issues/7397#issuecomment-459553803)

If subnet_mapping is used without setting the allocation ids, then the
state will be saved as if the configuration would be using the `subnets`
attribute instead of `subnet_mapping`. Thus, this ensures that if the
configuration is using `subnet_mapping` to define subnets, that should
be persisted in the state.

Notes:

See that if the load balancer uses static IP addresses, AWS requires to
set static IP addresses for every subnet attached to it. This means that
if the allocation id is set for one subnet, aws requires to set them for
all subnets. Thus, we have 3 scenarios:

1- Configure the LB using `subnets`.
2- Configure the LB using `subnet_mapping` to use static IP addresses.
3- Configure the LB using `subnet_mapping` without static IP addresses.

The 3th option only happens if the LB is inside a module that configures
the `subnet_mapping` with static IP addresses as options.