Update example linked role in Elasticsearch docs
Right now, an example in the docs for the `elasticsearch_domain` resource includes an `aws_iam_service_linked_role` with a service name of `elasticsearch.amazonaws.com`. This isn't the right domain – it's `es.amazonaws.com` – and running the code against AWS causes it to throw an error.

This updates that example to use the correct domain.