Add safe_delete option to resource/aws_sns_topic
`safe_delete` will protect against deletion of SNS topics if
there are other resources subscribed to that topic.

This can be a problem in the following scenario:
  - project A (with it's own terraform state) creates an `aws_sns_topic`
  - project B (with a different state) creates an
`aws_sns_topic_sbubscription`

Project A may not want to delete it's SNS topic if other projects
rely on that subscription. This `safe_delete` flag gives devs an way
to safely protect against this behavior.