RM-1456 Add & as valid in SG rule descriptions
Amazon documentation says that '&' is not a valid character in
Security Group rule descriptions, however the console allows using
'&'. The terraform-aws-provider has a validation to ensure that
descriptions match the documented allowed characters, which causes
errors when dealing with rules created outside of terraform that
include '&'.

This updates the Security Group rule description validator to allow
the use of '&'.