provider: Remove hardcoded AWS China handling
The usage of `IsChinaCloud()` (and previously `IsGovCloud()`) represented the first iteration of trying to support the 3 public AWS partitions when they offered differing capabilities due to running different service API versions. While this approach of checking the AWS partition provides an immediate fix for these situations, it does not automatically handle if/when those new capabilities are enabled.

In newer iterations of this handling, we attempt to simply ignore read errors based on specific error messages, like `UnsupportedOperation` and passthrough any create/update errors. See https://github.com/terraform-providers/terraform-provider-aws/pull/3794/files for an example of this.

Here we introduce the potentially breaking change of removing the AWS China protections in the aws_instance and aws_s3_bucket_object resources to further discourage the usage of this approach. Should any issues arise in AWS China from these removals, we will implement the error-based approach going forward. It is entirely possible that both EC2 Instance tagging on creation and S3 Object tagging is supported now in that partition though given both have been supported in AWS GovCloud for an extended period of time now.