Issue #376 AWS Opsworks Errors on 'apply'
Fixes #376

Changes proposed in this pull request:

* If Configuration Manager is Chef 12.2 and agent_version is LATEST (all
must match), don't add AgentVersion to the UpdateStack request.

Suggested fix for AgentVersion problem with OpsWorks Windows stacks.
Problem appears to be that AgentVersion gets read back from the AWS
stack (as LATEST) and re-sent in the UpdateStack request, but its
inclusion is always invalid in a Chef 12.2 stack.  This fix is designed
to lead to a meaningful error (ValidationException: Agent Version: can
only be set for Chef 11.10 and Chef 12) if the user specifies agent_version
as anything other than LATEST with a Chef 12.2 stack, so they're not left
with a specification that is ignored.  Not specifying agent_version at
all is still fine.