Refactor Lambda environment variable handling
This change refactors the environment parameter to accept a map of key/value pairs corresponding to environment variable name/values.

Previously, environment was a configuration block with one mandatory key, variables, which itself was a map of environment variable name/values. This set up, besides being somewhat redundant, also caused some problems for users attempting to pass empty maps to variables (see issues #1272 and #1110).

This refactor resolves the core problem in the issues mentioned above whereby users are unable to pass an empty map of variable to the lambda resource; setting environment to an empty map will create no environment variables as expected.

Fixes #1272 and #1110