route_table: fail creation if later read does not find resource
The typical route table creation steps involve making an ec2/CreateRouteTable
request, an ec2/DescribeRouteTables request, some other requests to update
the route table, followed by a final ec2.DescribeRouteTables request. If
the response to the final ec2/DescribeRouteTables request is
InvalidRouteTableID.NotFound, then the ID of the route table resource is
cleared so that the route table resource is not retained. However, the
creation is still considered a success. This results in either (1) failures
from other resources that depend upon the route table resource or (2) a report
of a successful apply even though one of the resources needed does not
actually exist.

These changes adjust the result of the creation in the case where the route
table cannot be found on the final ec2/DescribeRouteTables request. Instead
of reporting success for the creation, report failure.

Fixes #7644